refactor: Encapsulate ARCore session configuration
This commit refactors the `MainViewModel` to improve encapsulation by abstracting away the direct manipulation of the ARCore session configuration.

A new method, `updateAugmentedImageDatabase`, has been introduced in `ARCoreManager`. This method is now called from the `MainViewModel` to update the `AugmentedImageDatabase`, removing the need for the ViewModel to directly get, modify, and set the session's config.

Signed-off-by: Az <hereliesaz@gmail.com>